♻️ Refactor : 상세페이지 fetch함수들 커스텀 훅으로 변경(#10)
