+++ src/Components/Feed/Comment.jsx
@@ -1,28 +1,31 @@
 import React, { useState, useEffect } from 'react';
-import FetchComment from './FetchComment';
 import styled from 'styled-components';
 import elapsedTime from './elapsedTime';
 import { useNavigate } from 'react-router-dom';
 import { APIDefaultImage, profileImg } from './COMMON';
-
+import useFetchComment from './useFetchComment';
 const Comment = ({ feedList, commentList, setCommentList, isFetchData, setIsFetchData }) => {
-  useEffect(() => {
-    setIsFetchData(false); // 컴포넌트가 리렌더링될 때마다 데이터를 다시 불러오기 위해 isFetchData 상태를 false로 설정
-  }, []);
+  // useEffect(() => {
+  //   setIsFetchData(false); // 컴포넌트가 리렌더링될 때마다 데이터를 다시 불러오기 위해 isFetchData 상태를 false로 설정
+  // }, []);
   const navigate = useNavigate();
 
   function goProfile(item) {
     navigate('/myprofile', { state: item });
   }
+  // FetchDetailFeed 커스텀 훅 사용
+  const { getComment } = useFetchComment({
+    fetchType: 'comment',
+    postID: feedList.id,
+    setIsFetchData,
+    setCommentList,
+  });
+  useEffect(() => {
+    getComment(); // 컴포넌트가 마운트될 때 FetchDetailFeed 실행
+  }, []);
 
   return (
     <SCommentListLayout>
-      <FetchComment
-        fetchType="comment"
-        postID={feedList.id}
-        setIsFetchData={setIsFetchData}
-        setCommentList={setCommentList}
-      />
       {isFetchData ? (
         <div>
           {commentList.map(comment => {
